+++ packages/api/tests/helpers/create-test-data.js
@@ -290,12 +290,11 @@ async function createExecutionRecords({
 const generateListOfGranules = (count = 5000) => {
   const granuleRegString = () => randomStringFromRegex('^MOD09GQ\\.A[\\d]{7}\\.[\\w]{6}\\.006\\.[\\d]{13}$');
   const granuleLongString = () => randomStringFromRegex('^SWOT_L2_HR_Raster_\\_[\\w]{6}\\_[\\w]{1}\\_[\\d]{20}\\_[\\w]{6}\\_[\\d]{13}$');
-  const collectionId = () => `${randomString(3)}__${randomString(5)}`;
   const granules = [];
   const halfOfCount = count / 2;
   for (let i = 0; i < halfOfCount; i += 1) {
-    granules.push({ granuleId: granuleRegString(), collectionId: collectionId() });
-    granules.push({ granuleId: granuleLongString(), collectionId: collectionId() });
+    granules.push(granuleRegString());
+    granules.push(granuleLongString());
   }
   return granules;
 };