+++ packages/db/src/migrations/20240124101001_update_cumulus_id_add_indexes.sql
@@ -25,4 +25,3 @@ CREATE INDEX CONCURRENTLY IF NOT EXISTS granules_granule_id_index ON granules(gr
 SELECT CURRENT_TIMESTAMP;
 CREATE INDEX CONCURRENTLY IF NOT EXISTS granules_provider_collection_cumulus_id_granule_id_index ON granules(provider_cumulus_id, collection_cumulus_id, granule_id);
 SELECT CURRENT_TIMESTAMP;
-