+++ CHANGELOG.md
@@ -6,17 +6,42 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Breaking Changes
+
+- **CUMULUS-4473**
+  - Updated Granules Bulk Operations API endpoints to accept a list of granuleIds instead of
+    granule objects in the payload.
+  - Updated `/executions/search-by-granules` and `/executions/workflows-by-granules` endpoints
+    to accept granuleIds instead of granule objects in the payload.
+
+### Added
+
+- **CUMULUS-4473**
+  - Updated Granules Bulk Operations API endpoints to support `granuleInventoryReportName` and
+    `s3GranuleIdInputFile` in the payload.
+
+### Changed
+
+- **CUMULUS-4473**
+  - Updated Granules Bulk Operations return consistent output formats across different bulk opertions
+    (previously, some bulk operation aggregated errors while others returned per-granule errors)
+  - Removed the `getUniqueGranuleByGranuleId` and `getGranuleByUniqueColumns` functions from the
+    `@cumulus/db` package, since a single granule record can be retrieved using a unique `granule_id`.
+
+## [Unreleased]
+
 ### Migration Notes
 
 Please complete the following steps before upgrading Cumulus.
 
 #### CUMULUS-4459 New index added to the granules table to improve Dashboard performance
 
-- The fix introduced in CUMULUS-4459 requires a manual database update in the production environment. 
-  This step ensures the new index is created successfully, even in the unlikely event that the database-migration 
+- The fix introduced in CUMULUS-4459 requires a manual database update in the production environment.
+  This step ensures the new index is created successfully, even in the unlikely event that the database-migration
   Lambda function did not complete the index creation before timing out.
 
   Please follow the standard procedures for running a production database migration, and execute the following SQL to create the index:
+
   ```text
   CREATE INDEX CONCURRENTLY IF NOT EXISTS granules_collection_updated_idx ON granules (collection_cumulus_id, updated_at);
   ```
@@ -26,7 +51,7 @@ Please complete the following steps before upgrading Cumulus.
 - **CUMULUS-4459**
   - Added new index to the granules table to improve Dashboard performance.
 - **CUMULUS-4446**
-  - Updated all node lambdas/Core build environments to utilize node v22
+  - Updated all node lambdas/Core build environments to utilize node v22.
   - Updated cma-js dependency to 2.4.0
 - **CUMULUS-3574**
   - Granule file writes are now atomic. Previously, some granule files could be written even if others failed;
@@ -36,6 +61,11 @@ Please complete the following steps before upgrading Cumulus.
     This enhancement enables DAACs to migrate their existing RDS deployments to Aurora while
     reusing their existing security group, ensuring compatibility with existing
     `data-persistence-tf` and `cumulus-tf` modules.
+- **CUMULUS-4473**
+  - Updated Granules Bulk Operations API endpoints to:
+    - Support `granuleInventoryReportName` and `s3GranuleIdInputFile` in the payload.
+    - Return consistent output formats across endpoints (previously, some endpoints aggregated errors
+      while others returned per-granule errors)
 
 ### Added
 
@@ -54,9 +84,18 @@ Please complete the following steps before upgrading Cumulus.
     By setting `db_log_min_duration_ms` to a positive value (in milliseconds) and `enabled_cloudwatch_logs_exports`
     to `["postgresql"]`, RDS will log and export any database queries that take longer than that threshold.
     The module also configures the required RDS extensions and parameters necessary for slow query instrumentation.
+- **CUMULUS-4382**
+  - Migrated the granule-invalidator task to the `tasks` directory as part of a coreification task in support of providing rolling archive functionality.
 
 ### Changed
 
+- **CSD-82**
+  - Updated `/workflows/list` endpoint to accept `countOnly`, `prefix`, `infix`, `fields`, `limit`, and `order` query string params
+- **CUMULUS-4374**
+  - Updated example python Lambdas to utilize `uv` as their package manager. This change removes references to
+    pipenv. Developers should migrate to using `uv` to manage python dependencies and virtual envs which may
+    require reinstalling python libraries. This change also updates the names of the example python task services
+    because of a deployment race condition. These services are only used for integration tests.
 - **CUMULUS-4387**
   - Updated linting scripts to include `ruff` and `mypy` and enable lint rules in repo level
   `pyproject.toml` file.
@@ -65,9 +104,29 @@ Please complete the following steps before upgrading Cumulus.
 - **CUMULUS-4430**
   - Updated GitHub Actions to run `ruff` linting on PRs.
   - Updated GitHub Actions to run `eslint`, `markdownlint`, and `npm-package-json-lint` on PRs.
+- **CUMULUS-4433**
+  - Adds pre-commit config and hooks to the repository. Developers are encouraged to install pre-commit and read
+  the [pre-commit setup docs](./docs/development/pre-commit-setup.md) to ensure they have the correct setup.
 - **CUMULUS-4438**
   - Made `min_capacity` and `max_capacity` configurable in example/rds-cluster-tf
   - Made `archive_api_users` configurable in example/cumulus-tf
+- **CSD-61**
+  - Updated writeGranuleFromApi() endpoint to allow createdAt and updatedAt fields to be null.
+- **CUMULUS-4436**
+  - Created new documentation files for language best practices `docs/development/python-best-practices.md` and `docs/development/typescript-best-practices.md`.
+  - Updated documentation file `docs/development/quality-and-coverage.md` to be more repo wide and reference language best practices.
+  - Updated `docs/adding-a-task.md` to include instructions and expectations when adding a task.
+- **OTHER**
+  - Corrected misspelling in README.md related to installing `uv`.
+  - Added override for `tar` in package.json.
+
+
+### Fixed
+
+- **CUMULUS-4486**
+  - Fixed a small bug with `rulesHelpers` in which `rule.rule.meta.allowProviderMismatchOnRuleFilter` was erroring due to
+    database validation errors to instead refer to `rule.meta.allowProviderMismatchOnRuleFilter`
+  - Added `allowProviderMismatchOnRuleFilter` to the `meta` field of `rules` in `/api/lib/schemas`s
 
 ## [v21.2.0] 2025-12-06
 
@@ -6713,6 +6772,10 @@ the [release page](https://github.com/nasa/cumulus/releases)
   - Added `@cumulus/api-client/reconciliationReports`
 - **CUMULUS-1999**
   - Updated `@cumulus/common/util.deprecate()` so that only a single deprecation notice is printed for each name/version combination
+- **CUMULUS-4112**
+  - Updated `serveUtils.addGranules` to include writing granule files.
+- **CSD-84**
+  - Update error logging in `@cumulus/packages/cmr-client/searchConcept` to redact the authorization header value.
 
 ### Fixed